+++ api/src/main/java/io/strimzi/api/kafka/model/listener/arraylistener/KafkaListenerType.java
@@ -12,7 +12,8 @@ public enum KafkaListenerType {
     ROUTE,
     LOADBALANCER,
     NODEPORT,
-    INGRESS;
+    INGRESS,
+    CLUSTER_IP;
 
     @JsonCreator
     public static KafkaListenerType forValue(String value) {
@@ -27,6 +28,8 @@ public enum KafkaListenerType {
                 return NODEPORT;
             case "ingress":
                 return INGRESS;
+            case "cluster-ip":
+                return CLUSTER_IP;
             default:
                 return null;
         }
@@ -45,6 +48,8 @@ public enum KafkaListenerType {
                 return "nodeport";
             case INGRESS:
                 return "ingress";
+            case CLUSTER_IP:
+                return "cluster-ip";
             default:
                 return null;
         }