+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/ListenersValidatorTest.java
@@ -264,13 +264,13 @@ public class ListenersValidatorTest {
                 "listener " + name + " cannot configure bootstrap.host because it is not Route ot Ingress based listener",
                 "listener " + name + " cannot configure bootstrap.loadBalancerIP because it is not LoadBalancer based listener",
                 "listener " + name + " cannot configure bootstrap.nodePort because it is not NodePort based listener",
-                "listener " + name + " cannot configure bootstrap.annotations because it is not LoadBalancer, NodePort, Route, or Ingress based listener",
-                "listener " + name + " cannot configure bootstrap.labels because it is not LoadBalancer, NodePort, Route, or Ingress based listener",
+                "listener " + name + " cannot configure bootstrap.annotations because it is not LoadBalancer, NodePort, Route, Ingress or ClusterIP based listener",
+                "listener " + name + " cannot configure bootstrap.labels because it is not LoadBalancer, NodePort, Route, Ingress or ClusterIP based listener",
                 "listener " + name + " cannot configure brokers[].host because it is not Route ot Ingress based listener",
                 "listener " + name + " cannot configure brokers[].loadBalancerIP because it is not LoadBalancer based listener",
                 "listener " + name + " cannot configure brokers[].nodePort because it is not NodePort based listener",
-                "listener " + name + " cannot configure brokers[].annotations because it is not LoadBalancer, NodePort, Route, or Ingress based listener",
-                "listener " + name + " cannot configure brokers[].labels because it is not LoadBalancer, NodePort, Route, or Ingress based listener",
+                "listener " + name + " cannot configure brokers[].annotations because it is not LoadBalancer, NodePort, Route, Ingress or ClusterIP based listener",
+                "listener " + name + " cannot configure brokers[].labels because it is not LoadBalancer, NodePort, Route, Ingress or ClusterIP based listener",
                 "listener " + name + " cannot configure ipFamilyPolicy because it is internal listener",
                 "listener " + name + " cannot configure ipFamilies because it is internal listener"
         );
@@ -660,6 +660,100 @@ public class ListenersValidatorTest {
                 .build());
 
         assertThat(ListenersValidator.validateAndGetErrorMessages(2, asList(listener)), hasSize(0));
+    }
+
+    @ParallelTest
+    public void testClusterIPListenerWithoutTls() {
+        String name = "clusterip";
+
+        GenericKafkaListener listener1 = new GenericKafkaListenerBuilder()
+                .withName(name)
+                .withPort(9092)
+                .withType(KafkaListenerType.CLUSTER_IP)
+                .withTls(false)
+                .withNewConfiguration()
+                .withBrokers(new GenericKafkaListenerConfigurationBrokerBuilder()
+                                .withBroker(0)
+                                .withAdvertisedHost("my-host")
+                                .withAdvertisedPort(12345)
+                                .build(),
+                        new GenericKafkaListenerConfigurationBrokerBuilder()
+                                .withBroker(1)
+                                .withAdvertisedHost("my-host")
+                                .withAdvertisedPort(12346)
+                                .build())
+                .endConfiguration()
+                .build();
+
+        List<GenericKafkaListener> listeners = asList(listener1);
+
+        assertThat(ListenersValidator.validateAndGetErrorMessages(2, listeners), hasSize(0));
+    }
+
+    @ParallelTest
+    public void testClusterIPListener() {
+        String name = "clusterip";
+
+        GenericKafkaListener listener1 = new GenericKafkaListenerBuilder()
+                .withName(name)
+                .withPort(9092)
+                .withType(KafkaListenerType.CLUSTER_IP)
+                .withTls(true)
+                .withNewConfiguration()
+                .withIngressClass("my-ingress")
+                .withUseServiceDnsDomain(true)
+                .withExternalTrafficPolicy(ExternalTrafficPolicy.LOCAL)
+                .withIpFamilyPolicy(IpFamilyPolicy.REQUIRE_DUAL_STACK)
+                .withIpFamilies(IpFamily.IPV4, IpFamily.IPV6)
+                .withPreferredNodePortAddressType(NodeAddressType.INTERNAL_DNS)
+                .withLoadBalancerSourceRanges(asList("10.0.0.0/8", "130.211.204.1/32"))
+                .withFinalizers(asList("service.kubernetes.io/load-balancer-cleanup"))
+                .withNewBootstrap()
+                .withAlternativeNames(asList("my-name-1", "my-name-2"))
+                .withLoadBalancerIP("130.211.204.1")
+                .withNodePort(32189)
+                .withHost("my-host")
+                .withAnnotations(Collections.singletonMap("dns-anno", "dns-value"))
+                .endBootstrap()
+                .withBrokers(new GenericKafkaListenerConfigurationBrokerBuilder()
+                                .withBroker(0)
+                                .withAdvertisedHost("advertised-host")
+                                .withAdvertisedPort(9092)
+                                .withLoadBalancerIP("130.211.204.1")
+                                .withNodePort(32189)
+                                .withHost("my-host")
+                                .withAnnotations(Collections.singletonMap("dns-anno", "dns-value"))
+                                .build(),
+                        new GenericKafkaListenerConfigurationBrokerBuilder()
+                                .withBroker(1)
+                                .withAdvertisedHost("advertised-host")
+                                .withAdvertisedPort(9092)
+                                .withLoadBalancerIP("130.211.204.1")
+                                .withNodePort(32189)
+                                .withHost("my-host")
+                                .withAnnotations(Collections.singletonMap("dns-anno", "dns-value"))
+                                .build())
+                .endConfiguration()
+                .build();
+
+        List<GenericKafkaListener> listeners = asList(listener1);
+
+        List<String> expectedErrors = asList(
+                "listener " + name + " cannot configure useServiceDnsDomain because it is not internal listener",
+                "listener " + name + " cannot configure externalTrafficPolicy because it is not LoadBalancer or NodePort based listener",
+                "listener " + name + " cannot configure loadBalancerSourceRanges because it is not LoadBalancer based listener",
+                "listener " + name + " cannot configure finalizers because it is not LoadBalancer based listener",
+                "listener " + name + " cannot configure preferredAddressType because it is not NodePort based listener",
+                "listener " + name + " cannot configure bootstrap.loadBalancerIP because it is not LoadBalancer based listener",
+                "listener " + name + " cannot configure bootstrap.nodePort because it is not NodePort based listener",
+                "listener " + name + " cannot configure brokers[].loadBalancerIP because it is not LoadBalancer based listener",
+                "listener " + name + " cannot configure brokers[].nodePort because it is not NodePort based listener",
+                "listener " + name + " cannot configure ingressClass because it is not Ingress based listener",
+                "listener " + name + " cannot configure bootstrap.host because it is not Route ot Ingress based listener",
+                "listener " + name + " cannot configure brokers[].host because it is not Route ot Ingress based listener"
+        );
+
+        assertThat(ListenersValidator.validateAndGetErrorMessages(2, listeners), containsInAnyOrder(expectedErrors.toArray()));
     }
 
     @ParallelTest